+++ packages/dialtone-vue2/common/mixins/index.js
@@ -21,7 +21,3 @@ export {
 export {
   default as DtSkeletonMixin,
 } from './skeleton';
-
-export {
-  default as DtLocalizationMixin,
-} from './localization';