+++ packages/dialtone-vue2/localization/index.js
@@ -12,26 +12,27 @@ import ruRU from './ru-RU.ftl?raw';
 import esLA from './es-LA.ftl?raw';
 
 const dialtoneNamespace = 'dialtone';
+const allowedLocales = {
+  ENGLISH: 'en-US',
+  CHINESE: 'zh-CN',
+  DUTCH: 'nl-NL',
+  FRENCH: 'fr-FR',
+  GERMAN: 'de-DE',
+  ITALIAN: 'it-IT',
+  JAPANESE: 'ja-JP',
+  PORTUGUESE: 'pt-BR',
+  RUSSIAN: 'ru-RU',
+  SPANISH: 'es-LA',
+};
+const fallbackLocale = 'en-US';
 
 export class DialtoneLocalization {
-  constructor (locale = 'en-US') {
+  constructor (locale = null) {
     if (typeof DialtoneLocalization.instance === 'object') {
       return DialtoneLocalization.instance;
     }
 
-    this._locale = locale;
-    this._allowedLocales = {
-      ENGLISH: 'en-US',
-      CHINESE: 'zh-CN',
-      DUTCH: 'nl-NL',
-      FRENCH: 'fr-FR',
-      GERMAN: 'de-DE',
-      ITALIAN: 'it-IT',
-      JAPANESE: 'ja-JP',
-      PORTUGUESE: 'pt-BR',
-      RUSSIAN: 'ru-RU',
-      SPANISH: 'es-LA',
-    };
+    this._locale = locale || DialtoneLocalization.getPreferredLocale();
 
     const bundleSource = new RawBundleSource({
       resources: RawBundleSource.builtResources([
@@ -50,16 +51,48 @@ export class DialtoneLocalization {
 
     const localeManager = new LocaleManager({
       bundleSource,
-      allowedLocales: Object.values(this.allowedLocales),
-      fallbackLocale: locale,
-      preferredLocale: locale,
+      allowedLocales: Object.values(allowedLocales),
+      fallbackLocale,
+      preferredLocale: this._locale,
       namespaces: [dialtoneNamespace],
     });
 
     localeManager.install(dialtoneNamespace);
 
     DialtoneLocalization.instance = this;
+
+    if (typeof window !== 'undefined') {
+      /**
+       * @description
+       * When the browser storage changes, update the current locale
+       * @param event
+       */
+      window.onstorage = (event) => {
+        if (event.key === 'user-locale') {
+          this.currentLocale = event.newValue;
+        }
+      };
+    }
+
     return this;
+  }
+
+  static getPreferredLocale () {
+    if (typeof window === 'undefined' || !window.localStorage) {
+      return fallbackLocale;
+    }
+
+    const localStorageLanguage = window.localStorage.getItem('user-locale');
+
+    // Get the first two letters of the navigator language and check if it's in the allowed locales
+    const navigatorLanguage = Object.values(allowedLocales)
+      .find(locale => locale.startsWith(navigator.language.slice(0, 2)));
+
+    return localStorageLanguage || navigatorLanguage || fallbackLocale;
+  }
+
+  static getAllowedLocales () {
+    return allowedLocales;
   }
 
   $t (...args) {
@@ -76,15 +109,11 @@ export class DialtoneLocalization {
 
   set currentLocale (newLocale) {
     if (newLocale === this._locale) return;
-    if (!Object.values(this.allowedLocales).includes(newLocale)) {
-      throw new Error(`Locale ${newLocale} is not allowed, please use one of the following: ${Object.values(this.allowedLocales).join(', ')}`);
+    if (!Object.values(allowedLocales).includes(newLocale)) {
+      throw new Error(`Locale ${newLocale} is not allowed, please use one of the following: ${Object.values(allowedLocales).join(', ')}`);
     }
 
     this._locale = newLocale;
     useI18N(dialtoneNamespace).setI18N({ preferredLocale: newLocale }, dialtoneNamespace);
-  }
-
-  get allowedLocales () {
-    return this._allowedLocales;
   }
 }