+++ packages/dialtone-vue2/recipes/buttons/callbar_button_with_dropdown/callbar_button_with_dropdown.vue
@@ -74,14 +74,12 @@ import { DtDropdown } from '@/components/dropdown';
 import { DtIconChevronUp } from '@dialpad/dialtone-icons/vue2';
 import { DtRecipeCallbarButton, CALLBAR_BUTTON_VALID_WIDTH_SIZE } from '../callbar_button';
 import utils, { warnIfUnmounted } from '@/common/utils';
-import { DtLocalizationMixin } from '@/common/mixins';
+import { DialtoneLocalization } from '@/localization';
 
 export default {
   name: 'DtRecipeCallbarButtonWithDropdown',
 
   components: { DtRecipeCallbarButton, DtDropdown, DtButton, DtIconChevronUp },
-
-  mixins: [DtLocalizationMixin],
 
   /* inheritAttrs: false is generally an option we want to set on library
    components. This allows any attributes passed in that are not recognized
@@ -268,6 +266,7 @@ export default {
   data () {
     return {
       open: false,
+      i18n: new DialtoneLocalization(),
     };
   },