+++ packages/dialtone-vue3/localization/localization_default.story.vue
@@ -5,6 +5,7 @@ import { DialtoneLocalization } from '@/localization';
 
 export default {
   name: 'LocalizationDefault',
+
   components: {
     DtButton,
     DtStack,